MAIL finalization
Added MLton.Finalize, which implements finalization.  The only
function there currently is

	val MLton.Finalize.finalize: 'a * (unit -> unit) -> unit

finalize (x, f) causes f () to be run when x becomes unreachable. The
finalizer runs asynchronously in a separate thread after a garbage
collection determines that x is unreachable, which is done by
keeping a weak pointer to x.

The implementation works by treating GC as causing a signal, and
having a special signal handler for handling the GC signal.  So, the
finalizer runs in the usual signal handler thread.  This approach
required no new primitives, and only one new field in gcState,
gcSignalIsPending, to keep track of when a GC has been done and the
GC signal needs to be sent.

For now the GC signal is kept internal to the basis library
implementation and is not exposed to the user.