MAIL C codegen cleanup
Fixed bug that Mike Thomas found with _Thread_returnToC begin
undefined on Cygwin.  It was a problem with a missing "_" in
x86-main.h.

Moved the implementation of Array_length and Vector_length from the
codegens to the backend.

Starting moving the implementation of Runtime.GCField from the
codegens to the backend.  It works with the C codegen, but the native
codegen needs to be fixed to handle the new kinds of operands (offsets
of GCState).  For now, the backend switches on which codegen is being
used and generates the appropriate operand.  Once the native codegen
is updated, we should eliminate the switches and the
Machine.Operand.Runtime variant.

Added a new Runtime.GCField, ExnStack, and a new field to GC_state,
exnStack, which stores the current exception stack value.  This made
it easy to avoid complexities with accessing
s->currentThread->exnStack.

Reorganized the include directory to make it more clear what gets
included for each codegen.  Cleaned up the include file (now called
c-chunk.h) that goes in every file containing a C chunk generated by
the C codegen.  This file includes a lot less than it used to.  This
is in conjunction with turning on the code in the C codegen to output
the prototypes for all FFI functions that are called.  What this means
is that now when compiling -native false with FFI calls, there is no
need to include any other files in the C files being compiled, which
cleans up a long-standing problem.

Eventually, I'd like to move toward eliminating all knowledge in the
codegens about ExnStack, Frontier, StackBottom, and StackTop (except
for possibly some optimization hints).  Looking at c-chunk.h, there
are only a few uses left, corresponding to stack slots, object
allocation, stack push/pop, and return/raise.