MAIL fixed type inference bugs
Fixed two bugs in type inference that could cause the compiler to
raise the TypeError exception, along with a lot of XML IL.  The
type-check.sml regression contains simple examples of what failed.

The problem that Ken saw was that there was a type variable in scope
that did not appear in the type environment.  Hence, it was mistakenly
generalized over.  The fix was to keep all type variables in scope
also in the type environment.

The second problem, unrelated to Ken's, but that I noticed when
reading the code, was that bound type variables were not kept if they
did not occur in the type of the variable at which they were bound.
This was incorrect because they might be used in some inner scope.

Another thing I noticed and fixed was an omission in the XML type
checker, which did not check that the argument types of exceptions
were in scope.