MAIL fixed shrinker bug
Fixed bug in simplifier that could cause the following message

  shrinker raised Prim.apply raised assertion failure: SmallIntInf.fromWord

The problem only could have occurred in an SML/NJ-compiled MLton or in
one where asserts had been turned on.  The bug could conceivably have
happened if the shrinker visited a piece of dead code where
SmallIntInf.fromWord got an invalid word (or if there was some other
more horrible bug caused this in live code).