Fixed bug in the useless analysis that could cause the error message
	Type error: Type.equals
	{from = char vector, to = unit vector}
	Type error: analyze raised exception loopStatement: ...
	unhandled exception: TypeError

The bug happened because the useless analysis allowed a constant
string, which is of type char vector, to have its elements become
useless, requiring it to become a unit vector.  But the string
constant was still of type char vector.  For the fix, I could have
either replaced the string with a unit vector of appropriate size, or
weakened the analysis to mark the whole string useful.  I chose the
latter.

Added test useless-string.sml which catches the bug in older MLtons.