MAIL Const.SmallIntInf and IntInf.rep
Fixed bug that Henry saw (and Matthew found) with small negative
IntInf constants.  The problem was int Const.SmallIntInf.t, which
treated the word corresponding to a small IntInf inconsistently.
SmallIntInf.toWord shifted and added the tag bit, whit
SmallIntInf.fromWord assumed the tag bit wasn't there.  As a
consequence, the SSA simplifier treated small negative IntInf's as
large ones, i.e. as word vectors, and they got propagated to the x86
codegen which tried to treat an constant integer as a base address.

Unfortunately, it's not a good fix to use MLton.IntInf.rep, for two
reasons.  First, MLton still needs work when compiled using
mlton-stubs, so this just shifts the duplicated code from
Const.SmallIntInf to the stubs.  Second, even worse,
MLton.IntInf.Small w leaves the word untagged instead of tagged.  So
there wasn't much to be gained by moving the code to the stubs, since
we still need a slightly different treatment.  It might be worth
moving to the stubs if we change the meaning of MLton.IntInf.rep.
Anyone have any thoughts one way or the other on that?  I prefer
leaving it as is for two reasons.  One, that's the way it is now and I
assume at least PolySpace is using it.  Two, the shift and tag is
really part of the low level implementation and it's not clear it
should be exposed.

Hmmm.  It occurs to me that the right thing to do is to change rep
from:
       datatype rep =
 	  Small of word
 	| Big of word vector
to:
       datatype rep =
 	  Small of int
 	| Big of word vector

That has nothing to do with the bug at hand, but it does clarify the
issue with the tag, since it makes clear that we're dealing with an
int and not some low level representation.  It also solves the problem
with legacy code, since it will cause type errors for those who use
it.