MAIL writing profiling data when out of memory
Added fix to profiling so that the mlmon.out file is written even when
the program terminates due to running out of memory.  The fix is to
use the C atexit function to write out the profiling data (like we
used to ages ago) if for some reason the SML code didn't get to.