MAIL _export bugfix and improvements
Fixed the bug Ken Larsen triggered with mgtk.  It was due to broken
use of atomic{Begin,End} in MLtonThread.sml.  The reason the bug
didn't get caught in testing is that there is an optimization in the
basis library that turns atomic{Begin,End} into noops if signals
aren't handled by the program (because there can be no preemptive
thread switching).  The only testing I had done with _export had been
with a single-threaded program and hence the critical sections hadn't
been tested at all.  But mgtk uses MLton.Finalizable, which uses
signals, which triggered the bug.

First off, I decided to eliminate that optimization.  Partly because
of this problem, and partly because there are various places in the
runtime and basis where you want to do asserts on canHandle, but these
won't be right if atomic{Begin,End} are turned into noops.

I moved the implementation of MLton.FFI.register into MLton.Thread,
since it was the only user of MLton.Thread.setCallFromCHandler.  That
cleaned things up some.  Modulo that change, the bugfix was to insert
an extra atomicBegin in the C code that switches to the handler, as
well as an extra atomicBegin in the SML code to counteract an
atomicEnd that happens the first time the handler is started.

Added option

	-export-header {false|true}   output header file for _export's

Now, the header file for _exports is not created automatically.
Instead, you must call mlton with -export-header true, which will
output the header to stdout and exit.  This way, if you are creating a
library with _exports, you can create the header once and for all, and
users of your library don't have to know about it.

Added an example of mutually recursive C and SML functions to
doc/examples/ffi/export.sml.

Added support to allow values of type MLton.pointer to be passed
between C and SML.