MAIL profiler improvements
Added improvements discussed with Joe Hurd back in April.  The idea is
to shift -profile-split from compile time to mlprof time to allow much
faster user experimentation with splits and to create more precise
call graph information.  This is done by splitting everything at
compile time and keeping two sets of profiling data at run time, one
for the master versions of functions and one for the split copies.
Then, mlprof uses the split copies to build a precise call graph, and
merges nodes at the last possible minute, after the -keep expression
has been evaluated.

More specifically:
   Eliminated mlton -profile-split
   Added mlprof -split
   Changed mlprof -graph to -keep
   Eliminated mlprof -ignore

Now, the behavior of mlprof -keep is
1. Build the precise call graph using all the split copies of
   functions.  Functions with only one version have one split copy.
2. Evaluate the set of nodes corresponding to -keep.  The only
   trickiness is in evaluating thresh:

      (thresh x) = { f | f is -split and the ticks for this split
				version of f >= x }
                   U { f | f is not -split and the ticks for the
				unsplit version of f >= x }
   The rule for thresh-gc and thresh-stack is similar
3. For functions that are not -split, merge the split nodes.
4. Add a dotted edge from A to B if there is a path in the precise
   call graph from A to B going through nodes that don't correspond to
   a node in the merged graph.

Changed the behavior of -keep so that it controls which functions are
displayed in the table too.

Changed the behavior of -thresh x, which is now an abbreviation for
-keep '(thresh x)'.

Added some examples of lexical nesting to the profiling section of the
user guide.

Updated the mlprof man page.