MAIL handling types in C header for _exports
I went with the solution of textually copying types.h at compile time
into the exported header file.  That means that the file is completely
self-contained, but does leave around the typedefs.  So, C code that
#includes the header will have to know this.  That doesn't seem so
bad, since it has to know the types it dealing with when it calls the
exported functions anyway.