MAIL fixed bug in w64align
The return type of w64align was uint, not W64, which meant that it
could silently overflow.  This caused GC_arrayAllocate to mistakenly
continue, instead of reporting an out of memory error.

This fixes the bug that Matthew saw with

val a = Int64Array.tabulate (valOf Int.maxInt, fn i => Int64.fromInt i)