MAIL attributes in _import and _export expressions
Renamed _ffi as _import.  Left around _ffi for now.  We'll remove it
in a release or two.

Added attributes to _import and _export declarations.  These now look
like:

 	_import "name" <attribute>*: <ty>;
 	_export "name" <attribute>*: <ty>;

Here are the allowed attributes.

	<attribute> ::= cdecl | stdcall

cdecl, which we have always used, is the default, and means to use the
default C calling convention.  stdcall is only meaningful on
Cygwin/Windows and means to use the stdcall convention associated with
the Windows API and other libraries, like OpenGL.

Added checking for appropriate C types to _import.  We used to only
check _export.  This uncovered a couple of places in the basis library
where we pass types that are not allowed according to the MLton User
Guide: namely thread, prethread, and char vector array.  For now, I've
added a hack to the elaborator to allow these.

CFunctions now keep track of the calling convention (Cdecl or Stdcall)
and the argument types.

Moved the CFunction stuff from the backend to the frontend.  This is
actually a good thing, because it means that we will easily be able to
add source-level attributes for other things like "mayGC",
"modifiesFrontier", etc.  For now, MLton is conservative and assumes
that all C functions can modify the frontier and the stackTop and
mayGC (see elaborate-core.fun).  It's probably a bit much to add
attributes for all those things, and may even be a bit much that we
keep track of it within MLton.  An alternative approach would be to do
what Haskell does and to have "safe" and "unsafe" attributes, where
unsafe means that the C function doesn't do anything tricky like
modify the frontier, the stack, switch threads, or GC.  Hence a simple
and fast (and unsafe in general) call to C can be made.  safe is the
default.  Anyways, something to keep in mind for the future.

With the move, Prim.Name.FFI now takes a CFunction instead of a
string.  I alse created a new primitive name, FFI_Symbol, for imports
that aren't functions.

The x86 codegen mod was pretty easy -- I just turned off the stack pop
when the convention is stdcall.  Hopefully that doesn't mess up
anything.