MAIL fixed @MLton {fixed,max}-heap
Fixed @MLton max-heap, which was mistakenly ignored.  Cleaned up
@MLton fixed-heap.  Both fixed-heap and max-heap can use copying or
mark-compact collection.

Added check to runtime so that one can use max-heap or fixed-heap, but
not both.

The way both now work is that heapDesired size adjusts the amount of
heap space to meet the constraint.  Then, there is another check
before allocating a second semispace that the sum of the semispace
sizes meets the constraint.