MAIL Word64 is there
Word64 is implemented, much as Int64 is, with _import for all the
primitives.  This is especially bad for coercions that bounce through
LargeWord (which is now Word64), since they will involve a couple of C
calls and cannot be simplified away.

Hopefully the Word64 primitives can be added to the x86 codegen at the
same time as the Int64 ones.  As the primitives are added, they should
be propagated to basis-library/misc/primitive.sml, eliminated from
runtime/basis/Int/Word64.c, and uncommented in include/c-chunk.h.

Added a new regression, word-all.sml, that tests all of the functions
in all of the Word modules.