MAIL new front end
This checkin is the next phase in getting a proper front end for
MLton.  It does type checking of core SML programs correctly.  It
still does type checking after defunctorization and ignores types in
signature.  That will be fixed in the next phase.

I am pleased to report that the front end is quite fast.  For example,
it can lex, parse, and type check all of MLton (118K lines) in under
10 seconds on my 1.6GHz machine.

I am very interested in receiving bug reports and suggestions on how
to improve the front end.  In decreasing order of importance, I would
like to hear about:

1. Programs that are rejected but should be accepted.
2. Programs that are accepted but should be rejected.
3. Confusing type error messages.
4. Improvements to type error messages.

I would appreciate everyone starting to test this front end while
developing SML code.  If building MLton from the CVS is an impediment
to doing this, let me know and I will make an experimental release.

There is one kind of problem that I am aware of that will not be fixed
until the module-level checking is there.  MLton will now reject some
valid SML programs like the following.

	structure S:
	   sig
	      val f: 'a list -> 'a list
	   end =
	   struct
	      fun f _ = []
	   end

	val z = S.f [1, 2, 3]

The problem MLton has with this program is that the inferred type for
f is more general than the type given in the signature.  Then, at the
call to S.f, type inference is unable to deduce the result type.
Since the expression is expansive, MLton is unable to generalize the
free type variable, and reports an error.  I had to patch MLton in a
couple of places and a couple of the benchmarks and regressions to
work around this bug.


Here's an overview of what I did and how things now work.

All of the type inference code has been moved from operating on CoreML
to operating on Ast.   Here's how the front end now works.

     elaborate           convert
Ast -----------> CoreML ---------> XML

* Ast is as before, the raw, implicitly typed SML source.
* CoreML is changed and is now explicitly typed, polymorphic,
  direct-style, with nested patterns, and has no module-level
  constructs.
* XML is as before, explicitly typed, polymorphic, A-normal, and has
  flat patterns.

At the module level, the elaborate pass
* duplicates functors
* cuts structures according to signatures
* eliminates longids

Then within each structure or functor body, for each declaration
(<dec> in the SML grammar), the elaborate pass does three steps:
1. * type variable scope inference
2. * precedence parsing
   * _{ex,im}port expansion
   * profiling insertion
   * unification
3. * overloaded {constant, function, record pattern} resolution

Then, the convert pass does the following, all in a single step:
* linearization
* match compilation
* lookup constants
* polymorphic val dec expansion
* moves datatypes to toplevel

One consequence of doing overloading resolution on a per declaration
level instead of over the whole program means that some programs that
used to be accepted will now be rejected.  For example,

	val x = ref NONE
	structure S = struct end
	val _ = x := SOME 13

I did add a pass to the elaborator to combine declarations into as
large a block of <dec> as possible.  So, for example, the following
program is accepted.

	local
	   val r = ref NONE
	in
	   val _ = r := SOME 13
	end

Without that pass, the front end would treat this as a local <strdec>
instead of a local <dec>, and would doing overloading resolution on
the "val r" dec before the unificaiton for the "val _" dec, which
would result in an error.

Lookup constants is now a bit tricky.  The _build_const and _const
declarations are turned into thunks by the elaborator.  These thunks
are thawed during the conversion from CoreML to XML.  To build the
constants, we elaborate the basis library and convert to XML, using
the thunks to keep track of all the constants that are produced.
Then, we write these constants out the the C file.  When processing a
normal program, we use the thunks to lookup the constant in a hash
table produced from lib/<target>/constants.

Now, for a few other minor points.

Moved {compile,main}.sml to {compile,main}.fun to be more consistent.

Changed how equality is handled.  Instead of being magically added
in compile.fun, it is added via _prim in the basis library.

Took out Ast from Atoms.  Now, pretty printing of ILs is handled by
directly producing Layouts instead of first converting to Ast.

Removed the type information associated with primitives in
Prim.fun.

Changed the formatting of front-end error messages.

Made the naming of some type operators more consistent.

	dearray --> deArray
	dearrow --> deArrow
	deref --> deRef
	...