MAIL new front end
The next step towards a new front end: elaboration of signatures.

This involved adding in some old code that was ripped out a couple of
years ago to handle the representation of signatures.  This lives in
elaborate/interface.{fun.sig}.  The main trick is to use disjoint
sets to efficiently handle sharing of tycons and of structures and
then to copy signatures as dags rather than as trees.

There were a couple of fixes to make the basis library SML so that it
works with the new elaborator.  There was even one fix to MLton itself
-- in the x86 codegen there was a type definition in a signature that
was being shared.  SML/NJ allowed it, but the Definition (and hence
MLton) does not.

There is still one case that the signature elaborator doesn't get
right.  It is pretty obscure -- have a look at regression/where.sml if
you want to see it.

Signature elaboration has slowed down MLton type checking itself, but
not too badly --- to lex, parse, and type check MLton takes about 13s
on my machine.

There are still several things that need doing before the front end
meets the Definition.

1. Check that value declarations in structures match the
specifications in signatures.

2. Opaque matching needs to hide type information.

3. Functors need to be checked at the point of definition, instead of
at each application.

And of course, there still needs to be lots of work on error messages,
both for the new signature, as well as for the core.  Although I am
pretty much now developing exclusively with MLton type checking itself
and I am really liking the _ messages.