MAIL new front end
The next phase in the new front end: checking functors at the point of
definition.

This is implemented by building a dummy structure from the argument
signature and then applying the functor to the dummy strucure just as
we would for a real functor application.

Unfortunately, the fully-functorized programming style in MLton is
maximally bad for this method, because it causes each functor to be
checked many times.  For example, in the following code, the body of F
will be elaborated 4 times.

functor F () = ...
functor G () = ... F () ...
functor H () = ... G () ...
structure S = H ()

All this extra work means that elaboration of the MLton now takes
about 35 seconds.

I'll think about going to an approach where a functor summary (much
like in the Definition) is produced after elaboration of each functor,
so that the body need not be re elaborated.