MAIL new front end
The next phase in the new front end: matching value and constructor
specs in signatures.  It was pretty straightforward, requiring only
some unification.  Then only trickiness is that signature matching
needs to introduce value declarations to reparameterize values whose
type scheme changes.  For example, if we have in a structure

	val ('a, 'b) f: unit -> ('a list * 'b list) = fn () => ([], [])

and are matching that against the spec

	val f: unit -> ('c list * 'c list)

then signature matching has to introduce a new value that takes one
type argument ('c) and applies f to two type arguments ('c, 'c).

This checkin also fixes the bug that Jesper and Joe ran into that I
mentioned in a prior checkin.  So the following now works.

structure S:
   sig
      val f: 'a list -> 'a list
   end =
   struct
      fun f _ = []
   end
val z = S.f [1, 2, 3]


All that's left to get a viable new front end is:

* Opaque signature matches
* Speeding up functor checking with summaries