MAIL new front end
The next phase in the new front end: using functor summaries to speed
up checking of functor applications.

Now, when a functor is first type checked, we keep track of the dummy
argument structure and the dummy result structure, as well as all the
tycons that were created while elaborating the body.  Then, if we
later need to type check an application of the functor (as opposed to
defunctorize an application), we pair up tycons in the dummy argument
structure with the actual argument structure and then replace the
dummy tycons with the actual tycons in the dummy result structure,
yielding the actual result structure.  We also generate new tycons for
all the tycons that we created while originally elaborating the body.

With this improvement, type checking all of MLton has gone from 45s to
just under 20.  That's almost livable for day-to-day use.

All that's left is opaque matching (and lots of testing).