MAIL new front end
The next phase in the new front end: opaque signature constraints.
This is implemented by building a dummy structure realized from the
signature, just as we would for a functor argument when type checking
a functor.  The dummy structure contains exactly the type information
that is in the signature, which is what opacity requires.  We then
replace the variables (and constructors) in the dummy structure with
the corresponding variables (and constructors) from the actual
structure so that the translation to CoreML uses the right stuff.  For
each tycon in the dummy structure, we keep track of the corresponding
type structure in the actual structure.  This is used when producing
the CoreML types (see expandOpaque in type-env.{fun,sig})

Another way to look at things is that an opaque signature constraint
is equivalent to viewing the rest of the program as a functor
parameterized by a structure matching the signature being matched
opaquely.

The most annoying bit about all of this was getting the opacity of the
basis library right.  There were several problems.  The first stems
from free types in signatures.  For example, in BYTE we have

signature BYTE =
   sig
      ...
      val unpackString: Word8ArraySlice.slice -> string
   end

Here, Word8ArraySlice.slice is a free type.  Then, in BASIS_2002 we have

signature BASIS_2002 =
   sig
      ...
      structure Byte: BYTE
      ...
      structure Word8ArraySlice: MONO_ARRAY_SLICE
   end

The problem is that this signature establishes no connection between
the Word8ArraySlice.slice in Byte the slice Word8ArraySlice.  And we
cannot use a sharing constraint to make them the same, since the
Word8ArraySlice.slice in BYTE is not flexible.  So, we end up adding a
new where type to BASIS_2002

   where type Word8ArraySlice.slice = Word8ArraySlice.slice

This is a bit annoying, because it exposes the type of
Word8ArraySlice.slice, which is probably some record type.  To really
fix this, we need to add some opaque constraints earlier in the basis
library.

Another problem stems from the fact that we had the non-standard
structures (MLton, SMLofNJ, ...) outside of the opaque constraint on
Basis2002.  This meant that types that they used (e.g. MLton.Signal.t)
were not known to be the same as the basis library types
(e.g. Posix.Signal.signal).  One fix would be to add yet more where
constraints to BASIS_2002, but that would expose even more types.  So,
I decided a better fix was to move the non-standard structures into
BASIS_2002 and Basis2002.

I am beginning to wonder if using the :> for Basis2002 is the right
approach.  The problem is that this approach introduces lots of
potential bugs where we reject valid programs because two types are
not equal that should be, because we haven't added the appropriate
sharing or where.  If instead we were to use :, then we might accept
some programs because two types are equal that shouldn't be.  But that
seems more benign.  And when we learn it, we can use a :> somewhere
earlier in the basis library code to patch stuff up.  Although maybe
it's just a question of how quickly we can get to a correct BASIS_2002
so that the single big :> works.

I also think this would be an excellent time to drop support for
-basis 1997 and Basis1997.  Why, because it will be a whole lot of
additional work to get the opacity right there.  And, with the new
front end, it is much easier for people to migrate to the new basis.

Anyways, with all this, we now have a front end that works on all our
usual tests.  I am certain there are still many problems both in the
type checker and in missing sharing and wheres.  So, I am ready to
start getting some feedback.  Please start testing.  Soon, I will
revisit how type errors are displayed.