MAIL world no longer contains a preprocessed basis library
Moved the preprocessed basis library out of world.mlton (but not out
of the SML/NJ heap).  I did this because the world is starting to get
really huge with all of the type information and with the multiple
bases.  It's currently at about 25M.  Even with compression, it's
getting annoying having that in the package.

With all the basis library processing, compiling an empty file now
takes about 3.5 seconds on my 1.6 GHz machine, which I think is
livable.

In addition to making our binary packages much smaller, the even more
important advantage is that binary packages will now contain the basis
library SML source code.  So, it encourages people to look at it, and
allows people to easily fix bugs or make improvements.  It is also the
next step towards providing lots of other library sources in our
packages.

I would like to provide support for preprocessing and caching prefixes
of programs, but in a more general way that allows users to build a
world for any prefix, not just the basis library.  If any one else
wants to give this a try, it should be a pretty simple project that
will provide a nice feature.  Anyways, speak up on this list and we
can discuss designs.

Unfortunately, the change to main.fun in this checkin tickles a
long-latent bug in the SSA shrinker, which means that 20030716 fails
to compile this version of MLton.  I have checked in a fix which means
that this version of MLton can compile itself, but if you have any
earlier version before todays checkin, you will have problems.  The
way I worked around this was to compile a version of MLton with just
the SSA fix, and then to use that to compile this version.  A bit
messy.  I'm not sure what I'm gonna do to allow us to build on all our
platforms, which typically requires an older MLton to compile the
current one.