MAIL fixed Int64 and C codegen bugs
Fixed bugs with C codegen and Int64.  There were a couple of problems.
First, some primitives that were implemented by C routines were
treated as prims (not FFIs) by the C codegen.  As a consequence, there
was no C function prototype output in the chunk.  This fails when the
size of the return value is not an int, which is the case for many
Int64 primitives.  I fixed this problem by going through and changing
most _imports to _prims and then adding code to SsaToRssa that
carefully creates either the prim or the FFI call depending on which
codegen is used and the size of the operand.  That approach seems
better than what we had, because making things _prim (instead of
_import) in the basis library means that the SSA optimizer gets a
chance to improve things.  Also, the knowledge about what codegen
implements what primitive is all collected in SsaToRssa.

I made the C codegen use primitives for more Int64 ops.

There was a bug in how Int64 constants were printed.  To be safe, I
added a trailing "ll" to all of them.