MAIL _{ex,im}port and :>
_{ex,im}port now look inside opaque types to determine whether or not
a type can be used by the FFI.  So, for example, the following program
is accepted.

structure S:> sig type t end =
   struct type t = int end
val _ = _import "f": S.t -> S.t;

The decision was made to go this way, despite the seeming
contradiction with the opacity of :>, because there is no way to
specify in a signature that a type is "ffi-able".  So, it seems better
to allow the programmer the power to expose the property.  The
programmer can disallow it by wrapping a type in a datatype.