MAIL added MLton.Pointer
Added structure MLton.Pointer.  It has all the operations I proposed
in my earlier mail except for phantom types.  I didn't put those in
because they can now be done in user programs because the FFI looks
under opaque types.

Added get and set functions at all the types Int{8,16,32,64},
Real{32,64}, Word{8,16,32,64}.  I made them slightly more general than
before, treating the pointer as an array and allowing an array
offset.

	val getInt32: t * int -> Int32.int
	val setInt32: t * int * Int32.int -> unit

So, if in C we have "int *a" where a points to some array, then in SML
we can do

	val a = _import "a": MLton.Pointer.t;
	val _ = MLton.Pointer.getInt32 (a, 13)

You can follow a pointer by using offset 0.

Adding get* and set* required a family of new Pointer_ primitives in
the compiler.  These are implemented in SsaToRssa, which is when the
ArrayOffset operand is first introduced.

I went ahead and implemented MLton.Pointer.t as Word32.word, which
made it easy to do some stuff without adding more primitives.
Eliminating the built in pointer type also let me remove some code
from the compiler.  If we decide that we don't like having the
knowledge of the word size of pointers in the basis library, then it
will be easy enough to add a couple of primitives and the primitive
type.  Although if we do, I would recommend replacing the primitive
type in the front end with the appropriate word size instead of
pushing it all the way to the backend as we used to.

Eliminating the CPointer type from Rssa and Machine did create more
places for type unsafety in those ILs since we now must allow unsafe
pointer get and set on Word32s.  I don't know how much of a loss this
really is.