MAIL MLton.FFI now uses MLton.Pointer
Implemented the parameter passing for MLton.FFI using MLton.Pointer to
get/set array elements, thus eliminating the family of FFI primitives
that was used for array access.

Added MLton.Pointer.{get,set}Pointer, which allows one to get/set a
pointer.  The primitives are polymorphic so that we can use them to
access all kinds of values that are pointers (including arrays,
strings, and MLton.Pointer.t).  But for now, the signature only
exports them at MLton.Pointer.t.

One thing to keep in mind, and a mistake that I have now made twice
while working on the pointer stuff, is that a C array is *not* a
pointer.  That is, if in C you have

	int a[13];

Then it is *not* correct to do in SML

	val a = _import "a": MLton.Pointer.t;
        val _ = MLton.Pointer.getInt32 (a, 5)

Instead you must do in C

	int a[13];
	int *aPtr = &a;

followed by the SML

	val a = _import "aPtr": MLton.Pointer.t;
        val _ = MLton.Pointer.getInt32 (a, 5)