MAIL latex2html --> HeVeA
Switched the tool used to produce the HTML version of our user guide
from latex2html to hevea.  This was brought on by Debain bug #221341,
which pointed out that latex2html does not meet Debian's standard of
freeness, and hence if we continued using it we would have had to move
MLton out of their free section.

Anways, the change was relatively painless, and hevea certainly runs
quite a bit faster.