MAIL Added support for files larger than 2G
This included changing Position from Int32 to Int64.  It was
surprisingly easy to make the needed changes to the basis library.
The runtime was also pretty easy to update.  Now we compile with
-D_FILE_OFFSET_BITS=64 instead of -D_LARGEFILE64_SOURCE.  I added a
new typedef for Position as Int64, and updated the few routines that
need to use Position instead of Int (stat, lseek, ...).

Everything works on Linux.  I'll run tests on other platforms soon.

If people have some real apps that use 64 bit file offsets, I'd
appreciate some testing.