MAIL eliminated gcState.totalSwap
Eliminated gcState.totalSwap, which wasn't buying us much, since the
heap allocator backs off gracefully if it is unable to satisfy a
request, and since once it gets beyond ram it barely requests more
than the amount of live data anyways.  Also, implementing totalSwap
was causing portability problems.

Now, all that is left of setMemInfo is setting the RAM size.  I used
the sysconf call to get as much portability as possible there.